requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==16.3.0
falcon==1.1.0
gunicorn==19.6.0
jsl==0.2.4
jsonschema==2.5.1
pymysql==0.7.9
python-mimeparse==1.6.0   # via falcon
pyyaml==3.12
six==1.10.0               # via falcon, structlog
sqlalchemy==1.1.4
structlog==16.1.0
ujson==1.35
